fix: schema validation (#27)
* fix: JSON schema file path validation

Ensures paths with leading periods and/or slashes are invalid, and the path correctly includes/omits the file extension. The validation information is exposed via the filePath keyword.

* fix: validation of version

* fix: duplicate controllers

* feat: loosen restrictions on number of segments in action UUID

* fix: layout path validation

The layout $A0 is now valid, and $A2 is correctly invalid. When a file path, the .json extension must be included, and there must be no leading slash and/or period.

* feat: add custom error messages for patterns

---------

Co-authored-by: Richard Herman <geekyeggo@users.noreply.github.com>